remove contour ingressroute
