+++ source/store_test.go
@@ -99,11 +99,6 @@ func (suite *ByNamesTestSuite) TestAllInitialized() {
 	mockClientGenerator.On("DynamicKubernetesClient").Return(fakeDynamic.NewSimpleDynamicClientWithCustomListKinds(runtime.NewScheme(),
 		map[schema.GroupVersionResource]string{
 			{
-				Group:    "contour.heptio.com",
-				Version:  "v1beta1",
-				Resource: "ingressroutes",
-			}: "IngressRouteList",
-			{
 				Group:    "projectcontour.io",
 				Version:  "v1",
 				Resource: "httpproxies",
@@ -120,9 +115,9 @@ func (suite *ByNamesTestSuite) TestAllInitialized() {
 			}: "TCPIngressesList",
 		}), nil)
 
-	sources, err := ByNames(mockClientGenerator, []string{"service", "ingress", "istio-gateway", "contour-ingressroute", "contour-httpproxy", "kong-tcpingress", "fake"}, minimalConfig)
+	sources, err := ByNames(mockClientGenerator, []string{"service", "ingress", "istio-gateway", "contour-httpproxy", "kong-tcpingress", "fake"}, minimalConfig)
 	suite.NoError(err, "should not generate errors")
-	suite.Len(sources, 7, "should generate all six sources")
+	suite.Len(sources, 6, "should generate all six sources")
 }
 
 func (suite *ByNamesTestSuite) TestOnlyFake() {
@@ -157,9 +152,6 @@ func (suite *ByNamesTestSuite) TestKubeClientFails() {
 	_, err = ByNames(mockClientGenerator, []string{"istio-gateway"}, minimalConfig)
 	suite.Error(err, "should return an error if kubernetes client cannot be created")
 
-	_, err = ByNames(mockClientGenerator, []string{"contour-ingressroute"}, minimalConfig)
-	suite.Error(err, "should return an error if kubernetes client cannot be created")
-
 	_, err = ByNames(mockClientGenerator, []string{"kong-tcpingress"}, minimalConfig)
 	suite.Error(err, "should return an error if kubernetes client cannot be created")
 }
@@ -173,8 +165,6 @@ func (suite *ByNamesTestSuite) TestIstioClientFails() {
 	_, err := ByNames(mockClientGenerator, []string{"istio-gateway"}, minimalConfig)
 	suite.Error(err, "should return an error if istio client cannot be created")
 
-	_, err = ByNames(mockClientGenerator, []string{"contour-ingressroute"}, minimalConfig)
-	suite.Error(err, "should return an error if contour client cannot be created")
 	_, err = ByNames(mockClientGenerator, []string{"contour-httpproxy"}, minimalConfig)
 	suite.Error(err, "should return an error if contour client cannot be created")
 }
